Added mod routes. Cleaned up the GUI for mods.
