+++ models/ban.js
@@ -1,0 +1,38 @@
+const mongoose = require('mongoose');
+
+// SCHEMA SETUP
+const banSchema = new mongoose.Schema({
+    
+    ipban: Boolean,
+    userban: Boolean,
+
+    bannedPlayer: {
+        id: {
+            type: mongoose.Schema.Types.ObjectId,
+            ref: 'User',
+        },
+        username: String,
+        usernameLower: String,
+    },
+    
+    bannedIPs: [String],
+
+    modWhoBanned: {
+        id: {
+            type: mongoose.Schema.Types.ObjectId,
+            ref: 'User',
+        },
+        username: String,
+    },
+
+    whenMade: Date,
+    durationToBan: Date,
+    whenRelease: Date,
+
+    descriptionByMod: String,
+});
+
+// compile schema into a model
+const ban = mongoose.model('Ban', banSchema);
+
+module.exports = ban;
