+++ sockets/sockets.js
@@ -1754,108 +1754,6 @@ module.exports = function (io) {
         // when a user disconnects/leaves the whole website
         socket.on('disconnect', disconnect);
 
-        socket.on('modAction', async (data) => {
-            if (modsArray.indexOf(socket.request.user.username.toLowerCase()) !== -1) {
-                // var parsedData = JSON.parse(data);
-                const newModAction = {};
-                let userNotFound = false;
-
-                await data.forEach(async (item) => {
-                    if (item.name === 'banPlayerUsername') {
-                        // not case sensitive
-                        await User.findOne({ usernameLower: item.value.toLowerCase() }, (err, foundUser) => {
-                            if (err) { console.log(err); } else {
-                                // foundUser = foundUser[0];
-                                if (!foundUser) {
-                                    socket.emit('messageCommandReturnStr', { message: 'User not found. Please check spelling and caps.', classStr: 'server-text' });
-                                    userNotFound = true;
-                                    return;
-                                }
-                                // console.log(foundUser);
-                                newModAction.bannedPlayer = {};
-                                newModAction.bannedPlayer.id = foundUser._id;
-                                newModAction.bannedPlayer.username = foundUser.username;
-                                newModAction.bannedPlayer.usernameLower = foundUser.usernameLower;
-
-                                socket.emit('messageCommandReturnStr', { message: 'User found, Adding in details...\t', classStr: 'server-text' });
-                            }
-                        });
-                    } else if (item.name === 'typeofmodaction') {
-                        newModAction.type = item.value;
-                    } else if (item.name === 'reasonofmodaction') {
-                        newModAction.reason = item.value;
-                    } else if (item.name === 'durationofmodaction') {
-                        const oneSec = 1000;
-                        const oneMin = oneSec * 60;
-                        const oneHr = oneMin * 60;
-                        const oneDay = oneHr * 24;
-                        const oneMonth = oneDay * 30;
-                        const oneYear = oneMonth * 12;
-                        // 30 min, 3hr, 1 day, 3 day, 7 day, 1 month
-                        const durations = [
-                            oneMin * 30,
-                            oneHr * 3,
-                            oneDay,
-                            oneDay * 3,
-                            oneDay * 7,
-                            oneMonth,
-                            oneMonth * 6,
-                            oneYear,
-                            oneYear * 1000,
-                        ];
-                        newModAction.durationToBan = new Date(durations[item.value]);
-                    } else if (item.name === 'descriptionByMod') {
-                        newModAction.descriptionByMod = item.value;
-                    }
-                });
-
-                if (userNotFound === true) {
-                    return;
-                }
-
-                await User.findById(socket.request.user.id, (err, foundUser) => {
-                    if (err) { console.log(err); } else {
-                        newModAction.modWhoBanned = {};
-                        newModAction.modWhoBanned.id = foundUser._id;
-                        newModAction.modWhoBanned.username = foundUser.username;
-                    }
-                });
-
-                newModAction.whenMade = new Date();
-                newModAction.whenRelease = newModAction.whenMade.getTime() + newModAction.durationToBan.getTime();
-
-                setTimeout(() => {
-                    // console.log(newModAction);
-                    if (userNotFound === false && newModAction.bannedPlayer && newModAction.bannedPlayer.username) {
-                        modAction.create(newModAction, (err, newModActionCreated) => {
-                            if (newModActionCreated !== undefined) {
-                                // console.log(newModActionCreated);
-                                // push new mod action into the array of currently active ones loaded.
-                                currentModActions.push(newModActionCreated);
-                                // if theyre online
-                                if (newModActionCreated.type === 'ban' && allSockets[getIndexFromUsername(allSockets, newModActionCreated.bannedPlayer.username.toLowerCase(), true)]) {
-                                    allSockets[getIndexFromUsername(allSockets, newModActionCreated.bannedPlayer.username.toLowerCase(), true)].disconnect(true);
-                                } else if (newModActionCreated.type === 'mute' && allSockets[getIndexFromUsername(allSockets, newModActionCreated.bannedPlayer.username.toLowerCase(), true)]) {
-                                    allSockets[getIndexFromUsername(allSockets, newModActionCreated.bannedPlayer.username.toLowerCase(), true)].emit('muteNotification', newModActionCreated);
-                                }
-
-                                socket.emit('messageCommandReturnStr', { message: `${newModActionCreated.bannedPlayer.username} has received a ${newModActionCreated.type} modAction. Thank you :).`, classStr: 'server-text' });
-                            } else {
-                                socket.emit('messageCommandReturnStr', { message: 'Something went wrong...', classStr: 'server-text' });
-                            }
-                        });
-                    } else {
-                        let str = 'Something went wrong... Contact the admin! Details: ';
-                        str += `UserNotFound: ${userNotFound}`;
-                        str += `\t newModAction.bannedPlayer: ${newModAction.bannedPlayer}`;
-                        str += `\t newModAction.username: ${newModAction.username}`;
-                        socket.emit('messageCommandReturnStr', { message: str, classStr: 'server-text' });
-                    }
-                }, 3000);
-            } else {
-                // create a report. someone doing something bad.
-            }
-        });
 
         //= ======================================
         // COMMANDS
